refactor; continueScrolling and closeScroll move to ElasticSearchClient for reuse.
